[Maintenance] Move EmailChecker from Core into Behat
