+++ src/Sylius/Behat/Service/Checker/EmailCheckerInterface.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Component\Core\Test\Services;
+namespace Sylius\Behat\Service\Checker;
 
 interface EmailCheckerInterface
 {
